requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astroid==2.9.3
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
black==22.1.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coverage==6.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.4
executing==0.8.2
filelock==3.6.0
flake8==4.0.1
Flask==2.0.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.2
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy==0.931
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
packaging==21.3
Panda3D==1.10.10
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
py==1.11.0
pyarrow==7.0.0
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.12.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.3
pyzmq==22.3.0
requests==2.27.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
streamlit==1.7.0
terminado==0.13.1
testpath==0.6.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
virtualenv==20.13.1
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yapf==0.32.0
zipp==3.7.0
